requirements.txt
python-telegram-bot
pdf2docx
Pillow
python-dotenv
pdf2image
python-docx
xlsxwriter
pytesseract
